update import pool
