+++ src/pages/Swap/AssetAccordion.tsx
@@ -1,31 +1,18 @@
 import { useMemo } from 'react'
-import { Typography, Box, styled, useTheme } from '@mui/material'
+import { Typography, Box, useTheme } from '@mui/material'
 import Accordion from 'components/Accordion'
-import { useIsDarkMode } from 'state/user/hooks'
 import { useCallback, useState } from 'react'
 import { AllTokens } from 'models/allTokens'
 import CurrencyLogo from 'components/essential/CurrencyLogo'
+import Tag from 'components/Tag'
 
 export function AssetAccordion({ token }: { token?: AllTokens }) {
   const [expanded, setExpanded] = useState(false)
   const theme = useTheme()
-  const darkMode = useIsDarkMode()
 
   const handleChange = useCallback(() => {
     setExpanded(prev => !prev)
   }, [])
-
-  const Tag = styled(Box)(({ theme }) => ({
-    borderRadius: '10px',
-    boxShadow: '0px 3px 10px rgba(0,0,0,0.15)',
-    position: 'absolute',
-    right: 0,
-    top: 0,
-    fontSize: 12,
-    padding: '4px 12px',
-    background: darkMode ? '#484D50' : '#FFFFFF',
-    color: theme.palette.primary.main
-  }))
 
   const summary = useMemo(() => {
     return (
@@ -55,7 +42,7 @@ export function AssetAccordion({ token }: { token?: AllTokens }) {
           </Typography>
         </Box>
 
-        <Tag>{token && 'is1155' in token ? 'ERC1155' : 'ERC20'}</Tag>
+        <Tag sx={{ position: 'absolute', right: 0, top: 0 }}>{token && 'is1155' in token ? 'ERC1155' : 'ERC20'}</Tag>
       </Box>
     )
   }, [token])
